+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+master
+------
+
+* Added support for comments on paths, creating and listing comments. Delete comment added but API returns 404.
+* Added support for comments on content
+
 1.8.0
 -----